Update transcribe.py
Significantly updated transcribe.py to transcribe all files within a directory, compare the entire text of the original transcript to the entire predicted text, rather than trying to align time windows, and better handle silence detection. Still in the process of refining silence detection to be effective on all recordings.